Added c and start test including c
Added functions to compute c
	Added support functions BigLambdaJiJfFactor and AlignmentcFactor
	Added definition of c, including the J dep terms in c
	Added new def of the AngularCorrelationFactor to include c and b (necessary as well) in terms of spherical coordinates
	Added new def of the AngularCorrelationFactor in terms of vectors (in preparation for CRADLE main loop implementation)

Started tests to check inclusion of c
	Started test to verify the values of a, b, c and the auxiliary factors

+++ git/CRADLE++/include/CRADLE/Utilities.hh
@@ -223,7 +223,7 @@ namespace utilities {
   }
 
   inline double CalculateFierz(std::complex<double> cs, std::complex<double> csp, std::complex<double> ct, std::complex<double> ctp, std::complex<double> cv, 
-    std::complex<double> cvp, std::complex<double> ca, std::complex<double> cap, double mf, double mgt, double a, double b, double Z, double betaType)
+    std::complex<double> cvp, std::complex<double> ca, std::complex<double> cap, double mf, double mgt, double a, double b, int Z, int betaType)
   {
     if (std::isnan(a) && std::isnan(b)){
     double gamma = std::sqrt(1. - std::pow(FINESTRUCTURE * Z, 2.));
@@ -245,7 +245,7 @@ namespace utilities {
   }
 
   inline double CalculateBetaNeutrinoAsymmetry(std::complex<double> cs, std::complex<double> csp, std::complex<double> ct, std::complex<double> ctp, std::complex<double> cv, 
-    std::complex<double> cvp, std::complex<double> ca, std::complex<double> cap, double mf, double mgt, double a, double b, double energy, double Z, double betaType)
+    std::complex<double> cvp, std::complex<double> ca, std::complex<double> cap, double mf, double mgt, double a, double b, double energy, int Z, int betaType)
   {
     if (std::isnan(a) && std::isnan(b)){
     double coulombCorr = FINESTRUCTURE*Z/std::sqrt(1-EMASSC2*EMASSC2/energy/energy);
